* src/autologinpage.cpp, src/core/site.cpp, src/view/telnetcon.cpp, src/sitedlg.cpp: Add auto-login functionality. (CTelnetCon::ConnectThread) : Use select() in place of other blocking socket API without timeout.  This cause problems sometimes.
* src/generalprefpage.cpp,
src/view/telnetview.cpp,
src/dialog.cpp:
Little fix to accept NULL parent window.

* src/appconfig.cpp,
src/pcmanx_pure_gtk2.cpp:
Add socket timeout settings.

+++ src/generalprefpage.cpp
@@ -124,7 +124,7 @@ void CGeneralPrefPage::OnOK()
 	AppConfig.PopupNotifier =  gtk_toggle_button_get_active( 
 									GTK_TOGGLE_BUTTON(m_PopupNotifier));
 
-	AppConfig.PopupTimeout = gtk_spin_button_get_value( 
+	AppConfig.PopupTimeout = (int)gtk_spin_button_get_value( 
 									GTK_SPIN_BUTTON(m_PopupTimeout));
